// api/models/responses/AccountDetailsResponse.ts
export interface AccountDetailsResponse{
    responseCode: number,
    user: {
        id: number;
        name: string;
        email: string;
        title: "Mrs"| "Mr" | "Miss";
        birth_day: string;
        birth_month: string;
        birth_year: string;
        first_name: string;
        last_name: string;
        company: string;
        address1: string;
        address2: string;
        country: string;
        state: string;
        city: string;
        zipcode: string;
    }
}